Move `StringSplit` methods to StringConversion
